refactor(MatchingCandidates): break down 134-line component into modular structure (→35 lines)
- Extract useCandidates hook (28 lines): data fetching with loading/error states
- Extract LoadingSkeleton component (25 lines): reusable loading skeleton UI
- Extract CandidateCard component (71 lines): individual candidate display with animation
- Create index.js for clean re-export
- Main MatchingCandidates reduced from 134→35 lines
- Improved testability and reusability
- Follows established feature module pattern

Part of #203 (files over 200 lines reduction)
