+++ apps/registry/app/jobs/components/MatchingCandidates/LoadingSkeleton.jsx
@@ -1,0 +1,25 @@
+'use client';
+
+export const LoadingSkeleton = () => {
+  return (
+    <div className="mt-8">
+      <h2 className="text-2xl font-bold mb-4">Matching Candidates</h2>
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
+        {[...Array(4)].map((_, i) => (
+          <div
+            key={i}
+            className="animate-pulse bg-white rounded-lg shadow-sm p-4"
+          >
+            <div className="flex items-center space-x-4">
+              <div className="w-16 h-16 bg-gray-200 rounded-full"></div>
+              <div className="flex-1">
+                <div className="h-4 bg-gray-200 rounded w-3/4 mb-2"></div>
+                <div className="h-3 bg-gray-200 rounded w-1/2"></div>
+              </div>
+            </div>
+          </div>
+        ))}
+      </div>
+    </div>
+  );
+};
