+++ apps/registry/app/jobs/components/MatchingCandidates/index.js
@@ -1,0 +1,1 @@
+export { default } from './MatchingCandidates';
