+++ apps/registry/app/jobs/components/MatchingCandidates/useCandidates.js
@@ -1,0 +1,29 @@
+'use client';
+
+import { useState, useEffect } from 'react';
+import axios from 'axios';
+
+export const useCandidates = (jobId) => {
+  const [candidates, setcandidates] = useState([]);
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  useEffect(() => {
+    const fetchCandidates = async () => {
+      try {
+        const { data } = await axios.get(`/api/candidates?jobId=${jobId}`);
+        setcandidates(data.candidates);
+        setError(null);
+      } catch (err) {
+        console.error('Error fetching candidates:', err);
+        setError('Failed to load matching candidates');
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchCandidates();
+  }, [jobId]);
+
+  return { candidates, loading, error };
+};
